Refactoring async actions, adding sentry to track issues and fixing some erros
