+++ front/.env-example
@@ -6,3 +6,4 @@ SEND_TO=<Email to send report>
 EMAIL=<Email from>
 PASS=<Email from pass>
 FRESHBOOKS_API=><API Key>
+SENTRY_CODE=<SentryURL>
